proc->thread stage 5:  BUF/VFS clearance!  Remove the ucred argument from vop_close, vop_getattr, vop_fsync, and vop_createvobject.  These VOPs can be called from multiple contexts so the cred is fairly useless, and UFS ignorse it anyway.  For filesystems (like NFS) that sometimes need a cred we use proc0.p_ucred for now.
This removal also removed the need for a 'proc' reference in the related
VFS procedures, which greatly helps our proc->thread conversion.

bp->b_wcred and bp->b_rcred have also been removed, and for the same reason.
It makes no sense to have a particular cred when multiple users can
access a file.  This may create issues with certain types of NFS mounts
but if it does we will solve them in a way that doesn't pollute the
struct buf.