Misc interrupts/LWKT 1/2: interlock the idle thread.  Put execution of fast interrupts inside a critical section.  Make the hardclock and statclock INTR_FAST.  Implement the strict priority queue mechanism for LWKTs. Implement prioritized preemption for interrupt and softint preemption. Keep better stats.
Note: this commit hacks up the userland scheduler, in particular the
notion of 'curproc' because threaded interrupts really mess up the userland
scheduler's idea of curproc, which it uses to assume that the process is not
on a run queue even though it is runnable.  The next step will be to
separate out and cleanup the userland scheduler.