syscall messaging 2: Change the standard return value storage for system calls from proc->p_retval[] to the message structure embedded in the syscall. System calls used to set their non-error return value in p_retval[] but must now set it in the message structure.  This is a necessary precursor to any sort of asynchronizatino, for obvious reasons.
This work was particularly annoying because all the emualtion code declares
and manually fills in syscall argument structures.

This commit could potentially destabilize some of the emulation code but I
went through the most important Linux emulation code three times and tested it
with linux-mozilla, so I am fairly confident that I got it right.

Note: proper linux emulation requires setting the fallback elf brand to 3 or
it will default to SVR4.  It really ought to default to linux (3), not SVR4.

    sysctl -w kern.fallback_elf_brand=3