Add the 'wmake' script to /usr/bin and wmake support to Makefile.inc1.
This script allows you to build any element within /usr/src using a
buildworld environment and is extremely useful for debugging or
development work.  Instead of having to restart the buildworld
(even with -DNOCLEAN -DNOTOOLS it can take a while) this script can
be used to immediately build whatever piece of the world you are
working on.