Add the NO_KMEM_MAP kernel configuration option.  This is a temporary option that will allow developers to test kmem_map removal and also the upcoming (not this commit) slab allocator.  Currently this option removes kmem_map and causes the malloc and zalloc subsystems to use kernel_map exclusively.
Change gd_intr_nesting_level.  This variable is now only bumped while we
are in a FAST interrupt or processing an IPIQ message.  This variable is
not bumped while we are in a normal interrupt or software interrupt thread.

Add warning printf()s if malloc() and related functions detect attempts to
use them from within a FAST interrupt or IPIQ.

Remove references to the no-longer-used zalloci() and zfreei() functions.