A significant number of applications need access to kernel data structures.  Traditionally they have done a #define _KERNEL and then #include'd the files.  Unfortunately this can create huge conflicts between libc declarations and kernel declarations.  Also, every time we create a new dependancy between kernel includes we might windup causing another unexpected header file to be included into such a user program, causing its build to fail due to a conflict.
This commit introduces _KERNEL_STRUCTURES.  A user program which needs
access to kernel data structures sets this instead of _KERNEL.  This
define will cause the kernel data structure but not the inlines or
extern's to be brought in.

Only those header files which directly impact buildworld have been
modified so far, but the intent is for all significant kernel header files
to use this feature.