Check when M_PREPEND returns an empty mbuf.
This is just toe-dipping.  There are so many instances like this
across the source base that more work needs to be done on it.

I have added /* XXX-MBUF */ comments in places where I could not
add checks for empty mbufs; this is helpful for future work because
it is a potential error point.

It is better to put a comment than panic() messages because it can
be used when the mbuf allocator is rewritten or is taught how to
handle memory starvation situations.