Remove PUSER entirely.  Since p_priority has nothing to do with kernel thread priorities, the userland scheduler can now use the entire range 0-127.
Document the estcpu and other macros related to the p_priority calculation.

Adjust the default values such that a cpu bound nice -20 process will not
completely starve a nice +0 process (though it gets close!).  Note that a
nice +2 process will not run at all in the face of a cpu bound nice -20
process.