Deal with multicast packets in a manner similar to Solaris, RFC 3376, and draft-ietf-magma-msf-api-05.txt.   Multicast packets are not sent to multicast-unaware sockets or to sockets not bound to the interface the packet came in on.
The sysctl net.inet.udp.strict_mcast_mship enables the new requirements
by default.  Original operation may be recovered by setting the sysctl to 0.

There is a relatively minor scaling issue with the per-PCB membership
array scan, which is linear.  It is not a show stopper though.

Submitted by: "William A. Carrel" <william.a@carrel.org>
Reviewed by: Jeffrey Hsu <hsu@FreeBSD.org>, dillon