Implement variant symlinks!  In order to use variant symlinks you must first enable them with the vfs.varsym_enable sysctl.  Variant symlinks are disabled by default (the default will change in the future).  WORK TODO: system-wide and per-user variable sets need to be per-jail, and a method of listing existing variables is needed.
The 'varsym' utility may be used to set and retrieve system-wide or per-user
variables regardless of whether variant symlinks are enabled or not.  Note
that per-process variables may also be set by a program (e.g. like a csh
builtin or explicitly by a program).

When enabled, variable replacement will occur on softlinks which are
resolved as part of path lookup.  ${VARNAME} patterns within the softlink
will be replaced by the contents of the variable.  If a variable does not
exist, no replacement occurs.  ENAMETOOLONG will occur if a post-replacement
softlink pattern exceeds MAXPATHLEN (1024 bytes).

Note that variable names are case sensitive.  The per-process varsym list
is checked first, then the per-user list, then the system-wide list.