Add a new library, libcaps, which is intended to encompass DragonFly-specific core threading and IPC mechanisms.
As of this commit the library is very incomplete.  It works just well enough
to run the IPC test program unthreaded.  For the moment it is implemented
via unix domain sockets.  The IPC mechanism is eventually expected to be
moved directly into the kernel and use an upcall for replies.