Major GCC surgery.  Move most compiler-specific files into named subdirectories and do a major overhaul of the front-end binaries in /usr/bin.
    /usr/libdata/gcc -> /usr/libdata/gcc2
    /usr/libdata/ldscripts -> /usr/libdata/gcc2/ldscripts
    /usr/libexec/{all_gcc_bins} -> /usr/libexec/gcc2/*
    /usr/libexec/{elf,aout}/* -> /usr/libexec/gcc2/{elf,aout}/*

OBJFORMAT_PATH has been changed from a full path to a base path, because
the objformat binary (hardlinked to most compiler related binaries in
/usr/bin) must now interpret the command name and use one of two different
paths to find the actual binary.  Previously certain gcc binaries, such as
'cc' and 'cpp', were directly installed in /usr/bin.  That is no longer the
case.

Objformat now supports the GCCVER environment variable for vectoring the
compiler version.  This will be used to support gcc3 and other compilers.