It seems I missed a directory when doing the __FreeBSD__ -> __DragonFly__ conversion. This I believe makes world buildable with gcc3
NOTE:

    The things in the contrib areas will generate warnings about
__FreeBSD__ being redefined for the moment. To correct that
will require moving the files to where they are built from, and
correcting them there. Then patches will be sent to the owners of
the files, and finally we can update with corrected sources and
all will be back to normal. I will be going through these one by
one and moving the files that need to be fixed into place.
But I expect this to take some time.