isa_wrongintr() cannot depend on the (void *) unit argument pointing to a degenerate structure containing the irq, beacuse isa_wrongintr() is called only when the wrong APIC vector is used vs the one we thought we programmed.
This commit adjusts isa_wrongintr() to not depend on the unit argument,
fixing a crash that used to occur in the procedure.  It does not 'fix' the
interrupt routing bug responsible for causing the entry into isa_wrongintr(),
which is typically due to broken MP tables.  But at least the problem gets
reported.  FreeBSD-4.x just locks up without any error message in similar
situations.