Add patch infrastructure for contrib/ and similiar directories.
This allows the inclusion of patches in SRCS in the following form:
this,is,the,source.c.patch
This will create ${.OBJDIR}/this/is/the/source.c by applying above patch
to ${CONTRIBDIR}/this/is/the/source.c. The patched files are deleted on
"make clean" and created by "make depend".