Initial XIO implementation.  XIOs represent data through a list of VM pages rather then mapped KVM, allowing them to be passed between threads without having to worry about KVM mapping overheads, TLB invalidation, and so forth.
This initial implementation supports creating XIOs from user or kernel data
and copying from an XIO to a user or kernel buffer or a uio.  XIO are intended
to be used with CAPS, PIPES, VFS, DEV, and other I/O paths.

The XIO concept is an outgrowth of Alan Cox'es unique use of target-side
SF_BUF mapping to improve pipe performance.