Implement lwkt_abortmsg() support.  This function chases down a message and marks it MSGF_ABORTED, requeuing it to its target port for abort reprocessing. The chasedown may run through multiple message forwardings and even chase the message all the way back to the reply port (in which case the abort becomes a NOP).  The queueing of a replied message is delayed if there is an abort chasing it down until the abort chasedown has caught up to the message.
Support MSGF_PCATCH in the default waitport function.  If a signal is pending,
lwkt_default_waitport() (the default assigned to mp_waitport) will request
a message abort.  Note that we still have to wait for the message to be
returned after requesting a message abort.

Also do some minor performance tuning and adjust or move some of the inlines
from msgport2.h into lwkt_msgport.c.